requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Hugging Face
transformers==4.36.2
torch==2.1.2
sentencepiece==0.1.99

# Task Queue
celery==5.3.4
redis==5.0.1

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.26.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0

# Logging
structlog==24.1.0
